SeeFood/requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
gunicorn==20.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
